pyproject.toml
[project]
name = "notesproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.135.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt[crypto]>=2.12.1",
    "pyopenssl>=26.0.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
    "python-multipart>=0.0.22",
    "asyncpg>=0.31.0",
    "redis>=7.4.0",
    "alembic>=1.18.4",
    "starlette>=1.0.0",
    "uuid>=1.30",
]
